t2 f1

+++ 0x0C-more_malloc_free/2-calloc.c
@@ -24,7 +24,7 @@ ptr = malloc(size * nmemb);
 if (ptr == NULL)
 return (NULL);
 
-for (i = 0; i < nmemb; i++)
+for (i = 0; i < (int)nmemb; i++)
 ptr[i] = 0;
 return (ptr);
 }
